refactor(memory): trim legacy memory generation from brain/systems/memory
Deletes the dead legacy memory-generation modules (encoder, dedup, harvest,
integrity, lessons, narratives, repo_summary, retrieval, retrieval_pools,
scope, source_freshness, conflict_resolver) and three orphaned nightly
pipelines, moving still-live extraction/narrative behavior into
brain/systems/sessions. Keeps attention_controller, embeddings,
conflict_scout, truth_maintenance; leaves retrieval_feedback.py untouched
(owned by #346). Reconstructive memory is the durable replacement. No migration.

Closes #347
